+++ database/migrations/2020_05_30_121906_create_order_items_table.php
@@ -1,0 +1,42 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+use Shopper\Framework\Traits\Database\MigrationTrait;
+
+class CreateOrderItemsTable extends Migration
+{
+    use MigrationTrait;
+
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::create($this->getTableName('order_items'), function (Blueprint $table) {
+            $this->addCommonFields($table);
+
+            $table->string('name')->nullable()->comment('The product name at the moment of buying');
+            $table->string('sku')->nullable()->index();
+            $table->morphs('product');
+            $table->integer('quantity');
+            $table->decimal('price', 10, 2);
+
+            $this->addForeignKey($table, 'order_id', $this->getTableName('orders'));
+            $this->addForeignKey($table, 'shop_id', $this->getTableName('shops'), false);
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::dropIfExists($this->getTableName('order_items'));
+    }
+}