+++ src/Models/Order/Order.php
@@ -1,0 +1,116 @@
+<?php
+
+namespace Shopper\Framework\Models\Order;
+
+use Illuminate\Database\Eloquent\Model;
+use Shopper\Framework\Models\Address;
+use Shopper\Framework\Models\Shop\Shop;
+use Shopper\Framework\Models\User;
+
+class Order extends Model
+{
+    /**
+     * The attributes that are mass assignable.
+     *
+     * @var array
+     */
+    protected $fillable = [
+        'number',
+        'status',
+        'currency',
+        'shipping_total',
+        'shipping_method',
+        'notes',
+        'parent_order_id',
+        'shipping_address_id',
+        'shop_id',
+        'user_id',
+    ];
+
+    public function __construct(array $attributes = [])
+    {
+        // Set default status in case there was none given
+        if (!isset($attributes['status'])) {
+            $this->setDefaultOrderStatus();
+        }
+
+        parent::__construct($attributes);
+    }
+
+    /**
+     * Get the table associated with the model.
+     *
+     * @return string
+     */
+    public function getTable()
+    {
+        return shopper_table('orders');
+    }
+
+    /**
+     * Return total order.
+     *
+     * @return mixed
+     */
+    public function total()
+    {
+        return $this->items->sum('total');
+    }
+
+    /**
+     * Get the user Shipping Address for the order.
+     *
+     * @return \Illuminate\Database\Eloquent\Relations\BelongsTo
+     */
+    public function shippingAddress()
+    {
+        return $this->belongsTo(Address::class, 'shipping_address_id');
+    }
+
+    /**
+     * Return the associate User for this order.
+     *
+     * @return \Illuminate\Database\Eloquent\Relations\BelongsTo
+     */
+    public function customer()
+    {
+        return $this->belongsTo(User::class, 'user_id');
+    }
+
+    /**
+     * Get all items of the order.
+     *
+     * @return \Illuminate\Database\Eloquent\Relations\HasMany
+     */
+    public function items()
+    {
+        return $this->hasMany(OrderItem::class);
+    }
+
+    /**
+     * Get Order's shop instance.
+     *
+     * @return \Illuminate\Database\Eloquent\Relations\BelongsTo
+     */
+    public function shop()
+    {
+        return $this->belongsTo(Shop::class, 'shop_id');
+    }
+
+    /**
+     * Set a default value to an order.
+     *
+     * @return void
+     */
+    protected function setDefaultOrderStatus()
+    {
+        $this->setRawAttributes(
+            array_merge(
+                $this->attributes, [
+                    'status' => OrderStatus::PENDING,
+                ]
+            ),
+            true
+        );
+    }
+}